removed hardcoded

+++ delete_lambda_function.py
@@ -1,9 +1,17 @@
 import boto3
+import sys
+
+
+if len(sys.argv) < 2:
+    print("Usage: python script.py <function name>")
+    sys.exit(1)
+
+func_name = sys.argv[1]
 
 lambda_client = boto3.client('lambda', region_name='ap-southeast-1')
 
 response = lambda_client.delete_function(
-    FunctionName='SendSNSEmail'
+    FunctionName=func_name
 )
 
 print("Lambda function deleted.", response)
